backend/requirements.txt
Django==3.2
psycopg2==2.8.6
djangorestframework==3.12.4
requests==2.25.1
django-cors-headers==3.7.0
